Make Flows optional
